requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
et==0.0.2
idna==3.3
lxml==4.8.0
requests==2.27.1
SDK==1.0.0
topsdk==20.4.21.40
urllib3==1.26.9
